routes/index: Load summary data in the background
