+++ app/routes/index.js
@@ -1,10 +1,6 @@
 import Route from '@ember/routing/route';
-import { inject as service } from '@ember/service';
 
 export default Route.extend({
-
-    ajax: service(),
-
     headTags: [{
         type: 'meta',
         attrs: {
@@ -13,20 +9,7 @@ export default Route.extend({
         }
     }],
 
-    model() {
-        function addCrates(store, crates) {
-            for (let i = 0; i < crates.length; i++) {
-                crates[i] = store.push(store.normalize('crate', crates[i]));
-            }
-        }
-
-        return this.get('ajax').request('/api/v1/summary').then((data) => {
-            addCrates(this.store, data.new_crates);
-            addCrates(this.store, data.most_downloaded);
-            addCrates(this.store, data.just_updated);
-            addCrates(this.store, data.most_recently_downloaded);
-
-            return data;
-        });
-    }
+    setupController(controller) {
+        controller.get('dataTask').perform();
+    },
 });
