+++ app/templates/index.hbs
@@ -24,12 +24,12 @@
     <div id='stats'>
         <div class='downloads'>
             {{svg-jar "download"}}
-            <span class='num' data-test-total-downloads>{{format-num model.num_downloads}}</span>
+            <span class='num' data-test-total-downloads>{{if hasData (format-num model.num_downloads) "---,---,---"}}</span>
             <span class='desc small'>Downloads</span>
         </div>
         <div class='crates'>
             {{svg-jar "crate"}}
-            <span class='num' data-test-total-crates>{{format-num model.num_crates}}</span>
+            <span class='num' data-test-total-crates>{{if hasData (format-num model.num_crates) "---,---"}}</span>
             <span class='desc small'>Crates in stock</span>
         </div>
     </div>
@@ -37,27 +37,27 @@
 
 <div id='home-crates' class='crate-lists'>
     <div id='new-crates' data-test-new-crates>
-        <h2>New Crates</h2>
+        <h2>New Crates {{#if dataTask.isRunning}}<img src="/assets/ajax-loader.gif">{{/if}}</h2>
         {{crate-list crates=model.new_crates}}
     </div>
     <div id='most-downloaded' data-test-most-downloaded>
-        <h2>Most Downloaded</h2>
+        <h2>Most Downloaded {{#if dataTask.isRunning}}<img src="/assets/ajax-loader.gif">{{/if}}</h2>
         {{crate-list crates=model.most_downloaded}}
     </div>
     <div id='just-updated' data-test-just-updated>
-        <h2>Just Updated</h2>
+        <h2>Just Updated {{#if dataTask.isRunning}}<img src="/assets/ajax-loader.gif">{{/if}}</h2>
         {{crate-list crates=model.just_updated}}
     </div>
     <div id='most-recently-downloaded' data-test-most-recently-downloaded>
-        <h2>Most Recent Downloads</h2>
+        <h2>Most Recent Downloads {{#if dataTask.isRunning}}<img src="/assets/ajax-loader.gif">{{/if}}</h2>
         {{crate-list crates=model.most_recently_downloaded}}
     </div>
     <div id='keywords' data-test-keywords>
-        <h2>Popular Keywords {{#link-to 'keywords'}}(see all){{/link-to}}</h2>
+        <h2>Popular Keywords {{#link-to 'keywords'}}(see all){{/link-to}} {{#if dataTask.isRunning}}<img src="/assets/ajax-loader.gif">{{/if}}</h2>
         {{keyword-list keywords=model.popular_keywords}}
     </div>
     <div id='categories' data-test-categories>
-        <h2>Popular Categories {{#link-to 'categories'}}(see all){{/link-to}}</h2>
+        <h2>Popular Categories {{#link-to 'categories'}}(see all){{/link-to}} {{#if dataTask.isRunning}}<img src="/assets/ajax-loader.gif">{{/if}}</h2>
         {{category-list categories=model.popular_categories}}
     </div>
 </div>
